requirements.txt
torch
torchvision
matplotlib
numpy
librosa
inflect
scipy
Unidecode
pillow
tqdm
configargparse
